move rpc internals to package
