+++ packages/sdp-rpc/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "@sdp/rpc",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./src/index.ts",
+      "default": "./src/index.ts"
+    },
+    "./errors": {
+      "types": "./src/errors.ts",
+      "import": "./src/errors.ts",
+      "default": "./src/errors.ts"
+    },
+    "./relay": {
+      "types": "./src/relay.ts",
+      "import": "./src/relay.ts",
+      "default": "./src/relay.ts"
+    },
+    "./solana": {
+      "types": "./src/solana.ts",
+      "import": "./src/solana.ts",
+      "default": "./src/solana.ts"
+    }
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit",
+    "lint": "biome check src/"
+  },
+  "dependencies": {
+    "@sdp/types": "workspace:*",
+    "@solana/addresses": "catalog:",
+    "@solana/kit": "catalog:"
+  }
+}
